app-engine/requirements.txt
Flask==1.1.2
google-cloud-datastore==1.12.0
google-cloud-firestore
google-auth==1.16.0
requests==2.23.0
gunicorn==20.0.4
google-cloud-pubsub
google-cloud-storage
PyYAML
firebase-admin